Выполнен полный переход на openapi. Созданы и добавлены все Api

+++ bank-employees-salaries-impl/src/main/java/com/anybank/repository/GradeRepository.java
@@ -1,7 +1,8 @@
 package com.anybank.repository;
 
-import com.anybank.model.Grade;
+import com.egorov.model.Grade;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GradeRepository extends JpaRepository<Grade, Integer> {
+public interface GradeRepository extends JpaRepository<Grade, Long> {
+
 }
